Move EvFDaqDirector and FastMonitoringService in the shared library, not in plugins.
