+++ EventFilter/Utilities/plugins/SealModule.cc
@@ -2,16 +2,16 @@
 #include "FWCore/ServiceRegistry/interface/ServiceMaker.h"
 //#include "EventFilter/Utilities/interface/MicroStateService.h"
 //#include "EventFilter/Utilities/interface/TimeProfilerService.h"
+#include "EventFilter/Utilities/interface/EvFDaqDirector.h"
+#include "EventFilter/Utilities/interface/FastMonitoringService.h"
 #include "EventFilter/Utilities/plugins/ExceptionGenerator.h"
 #include "EventFilter/Utilities/plugins/EvFRecordInserter.h"
 #include "EventFilter/Utilities/plugins/EvFRecordUnpacker.h"
 #include "EventFilter/Utilities/plugins/EvFBuildingThrottle.h"
-#include "EventFilter/Utilities/plugins/EvFDaqDirector.h"
 #include "EventFilter/Utilities/plugins/RawEventFileWriterForBU.h"
 #include "EventFilter/Utilities/plugins/RecoEventWriterForFU.h"
 #include "EventFilter/Utilities/plugins/RecoEventOutputModuleForFU.h"
 #include "EventFilter/Utilities/plugins/RawEventOutputModuleForBU.h"
-#include "EventFilter/Utilities/plugins/FastMonitoringService.h"
 #include "EventFilter/Utilities/plugins/DaqFakeReader.h"
 
 using namespace edm::serviceregistry;
