+++ EventFilter/Utilities/src/DirManager.cc
@@ -1,4 +1,4 @@
-#include "DirManager.h"
+#include "EventFilter/Utilities/interface/DirManager.h"
 #include "FWCore/Utilities/interface/Exception.h"
 #include <iostream>
